autonomous-inventory-agent/requirements.txt
pandas
numpy
streamlit
groq
python-dotenv
simpy
ortools
matplotlib